+++ test/test_integration_pumactl.rb
@@ -108,6 +108,7 @@ class TestIntegrationPumactl < TestIntegration
     cli_pumactl "stop", unix: true
 
     _, status = Process.wait2(@pid)
+    assert_equal 0, status
     @server = nil
   end