+++ test/test_out_of_band_server.rb
@@ -60,8 +60,8 @@ class TestOutOfBandServer < Minitest::Test
     end
 
     @server = Puma::Server.new app, Puma::Events.strings, out_of_band: [oob], **options
-    @server.min_threads = options[:min_threads] || 1
-    @server.max_threads = options[:max_threads] || 1
+    @server.instance_variable_set :@min_threads, options[:min_threads] || 1
+    @server.instance_variable_set :@max_threads, options[:max_threads] || 1
     @port = (@server.add_tcp_listener '127.0.0.1', 0).addr[1]
     @server.run
   end