+++ History.md
@@ -8,6 +8,7 @@
 
 * Refactor
   * Change Events#ssl_error signature from (error, peeraddr, peercert) to (error, ssl_socket) (#2375)
+  * Remove unneeded attr_accessor statements from Server (#2373)
 
 ## 5.0.0